Blog/user/search_indexes.py
__data__ = '2019-04-04 18:28'
__author__ = 'Kai'

from haystack import indexes
from .models import Article


class ArticleIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, template_name=True)

    def get_model(self):
        return Article

    def index_queryset(self, using=None):
        return self.get_model().objects.all()

