Blog/requirements.txt
amqp==2.4.2
backports.csv==1.0.6
beautifulsoup4==4.7.1
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.9
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-redis==4.10.0
django-redis-sessions==0.6.1
django-reversion==3.0.3
django-simple-captcha==0.5.5
django-tinymce==2.8.0
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.17.1
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
jdcal==1.4
jieba==0.39
kombu==4.3.0
mutagen==1.42.0
odfpy==1.4.0
openpyxl==2.6.1
pbr==5.1.3
Pillow==5.4.1
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
PyYAML==3.13
redis==3.2.1
requests==2.21.0
six==1.12.0
soupsieve==1.8
sqlparse==0.2.4
stevedore==1.30.1
tablib==0.13.0
tornado==6.0.1
xlrd==1.2.0
xlwt==1.3.0
